+++ app/data/test-data.js
@@ -866,7 +866,7 @@ var allInOne =
       {"code": "c2", "text": "Answer 2", "other": null},
       {"code": "c3", "text": "Answer 3", "other": null},
       {"code": "c4", "text": "Other:", "other": true}],
-      "skipLogic": null, "editable": null, "defaultAnswer": 'Answer 2', "displayControl": null, "calculationMethod": null, "items": null},
+      "skipLogic": null, "editable": null, "defaultAnswer": {'text': 'Answer 2'}, "displayControl": null, "calculationMethod": null, "items": null},
     {"questionCode": "type10", "dataType": "CWE", "header": false, "units": null, "codingInstructions": null,
       "questionCardinality": null, "answerCardinality": null, "question": "With data type CWE", "answers": [
       {"label": "1", "code": "c01", "text": "With a label 1", "score": 1, "other": null},
@@ -1134,7 +1134,7 @@ var allInOne =
         }
       ],
       "items":[
-        {"questionCode": "colA", "question": "A ST", "dataType": "ST",
+        {"questionCode": "colA", "question": "A ST", "dataType": "ST", "prefix": "Pre. A:",
           "displayControl": {
             "colCSS": [{"name": "width", "value":"25%"},
                     {"name": "min-width", "value":"10%"}]
@@ -1238,6 +1238,14 @@ var allInOne =
         {"code": "c2", "text": "Answer 2", "other": null},
         {"code": "c3", "text": "Answer 3", "other": null},
         {"code": "c4", "text": "Other:", "other": true}]
+    },
+    // an item with prefix
+    {
+      "questionCode": "with_prefix",
+      "dataType": "ST",
+      "header": false,
+      "prefix": "Prefix A:",
+      "question": "Question display text"
     },
 
 
@@ -1380,7 +1388,7 @@ var formWithUserData =
       {"code": "c2", "text": "Answer 2", "other": null},
       {"code": "c3", "text": "Answer 3", "other": null},
       {"code": "c4", "text": "Other:", "other": true}],
-      "skipLogic": null, "editable": null, "defaultAnswer": 'Answer 2', "displayControl": null, "calculationMethod": null, "items": null},
+      "skipLogic": null, "editable": null, "defaultAnswer": {'text':'Answer 2'}, "displayControl": null, "calculationMethod": null, "items": null},
     // answer list, user selected value
     {"questionCode": "q6", "dataType": "CNE", "header": false, "units": null, "codingInstructions": null,
       "value": {"code": "c2", "text": "Answer 2", "other": null, label: null, score: null},
@@ -1416,15 +1424,15 @@ var formWithUserData =
       "skipLogic": null, "editable": null, "defaultAnswer": null, "displayControl": null, "calculationMethod": null, "items": null},
     // multi-select CWE, with a mix of selected answers and user typed data
     {"questionCode": "q9", "dataType": "CWE", "header": false, "units": null, "codingInstructions": null,
-      "value": [{"code": "c2", "text": "Answer 2"},{"text": "User created answer"}],
+      "value": [{"code": "c2", "text": "Answer 2"},"User created answer"],
       "questionCardinality": null, "answerCardinality": {"min": "0", "max": "*"}, "question": "Multi Selection on CWE", "answers": [
       {"code": "c1", "text": "Answer 1"},
       {"code": "c2", "text": "Answer 2"},
       {"code": "c3", "text": "Answer 3"},
       {"code": "c4", "text": "Answer 4"}],
       "skipLogic": null, "editable": null, "defaultAnswer": null, "displayControl": null, "calculationMethod": null, "items": null},
-    // multi-select CWE with a search autocomplter
-    {"questionCode": "q10", "dataType": "CWE", "header": false, "units": null, "codingInstructions": null, "value": [{"code": "c2", "text": "Answer 2"},{"text": "User created answer"}],
+    // multi-select CWE with a search autocompleter
+    {"questionCode": "q10", "dataType": "CWE", "header": false, "units": null, "codingInstructions": null, "value": [{"code": "c2", "text": "Answer 2"},"User created answer"],
       "questionCardinality": null, "answerCardinality": {"min": "0", "max": "*"}, "question": "Multi Selection on CWE, a search field",
       "externallyDefined":"https://clin-table-search.lhc.nlm.nih.gov/api/variants/v3/search?df=AlleleID,RefSeqID,GeneSymbol,NucleotideChange,AminoAcidChange&ef=RefSeqID,GeneSymbol,NucleotideChange,AminoAcidChange,phenotypes",
       "skipLogic": null, "editable": null, "defaultAnswer": null, "displayControl": null, "calculationMethod": null, "items": null},
@@ -1588,7 +1596,144 @@ var formWithUserData =
       unit: {name: "kgs"},
       units: [{name: "lbs", default: true},{name: "kgs"}],
       value: 456
-    }
+    },
+    // user value and default answers on CWE typed items when answers are displayed as radio buttons and checkboxes
+    {
+      "questionCode": "cwe-checkbox-user-value",
+      "question": "CHECKBOX --CWE, with user typed value",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "*"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "value": ["user typed value"]
+    },
+    {
+      "questionCode": "cwe-checkbox-user-value-and-answer-code",
+      "question": "CHECKBOX --CWE, with user typed value and an answer by code",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "*"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "value": ["user typed value", {"code": "c1"}]
+    },
+    {
+      "questionCode": "cwe-checkbox-user-value-and-answer-text",
+      "question": "CHECKBOX --CWE, with user typed value and an answer by text",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "*"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "value": ["user typed value", {"text": "Answer B"}]
+    },
+    {
+      "questionCode": "cwe-checkbox-user-value-and-answer",
+      "question": "CHECKBOX --CWE, with user typed value and an answer",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "*"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "value": ["user typed value", {"code": "c3", "text": "Answer C"}]
+    },
+
+    {
+      "questionCode": "cwe-checkbox-default-answer",
+      "question": "CHECKBOX --CWE, with off-list default answer",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "*"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "defaultAnswer": ["off-list default answer"]
+    },
+    {
+      "questionCode": "cwe-radio-user-value",
+      "question": "RADIO layout --CWE, with user typed value",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "1"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "value": "user typed value"
+    },
+    {
+      "questionCode": "cwe-radio-user-value-nodisplay",
+      "question": "RADIO layout --CWE, with user typed value, no-matching",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "1"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "value": {"text": "user typed value"} // no matching, not to be displayed
+    },
+    {
+      "questionCode": "cwe-radio-default-answer",
+      "question": "RADIO layout --CWE, with off-list default answer",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "1"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "defaultAnswer": "off-list default answer"
+    },
+    {
+      "questionCode": "cwe-radio-default-answer-nodisplay",
+      "question": "RADIO layout --CWE, with off-list default answer, no-matching",
+      "dataType": "CWE",
+      "answerCardinality": {"min": "0", "max": "1"},
+      "displayControl": {
+        "answerLayout": {"type":"RADIO_CHECKBOX", "columns": "2"}
+      },
+      "answers": [
+        {"code": "c1", "text": "Answer A"},
+        {"code": "c2", "text": "Answer B"},
+        {"code": "c3", "text": "Answer C"},
+      ],
+      "defaultAnswer": {"text": "off-list default answer"} // no matching, not to be displayed
+    },
 
   ]};
 
@@ -6345,7 +6490,7 @@ var defaultAnswerForm = {
       "defaultAnswer": {"code": "B"}
     },
     {
-      "question": "List via answer text (not in specification)",
+      "question": "CWE not-on-list value",
       "questionCode": "ansTextDefault",
       "questionCardinality": {
         "min": "1",