+++ test/karma/fhir-sdc.spec.js
@@ -12,7 +12,7 @@ function createLinkId(lfItems, codeList) {
   if(!codeList) {
     codeList = [];
   }
-  
+
   lfItems.forEach(function(item){
     codeList.push(item.questionCode);
     item.linkId = codeList.join('_');
@@ -361,6 +361,35 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
 
           });
 
+          it('should covert prefix of an item', function () {
+            var fhirData = {
+              title: 'test title',
+              name: 'test name',
+              version: '0.0.1',
+              resourceType: 'Questionnaire',
+              "meta": {
+                "profile": [
+                  "http://hl7.org/fhir/4.0/StructureDefinition/Questionnaire"
+                ]
+              },
+              status: 'draft',
+              item: [{
+                text: 'item a',
+                linkId: '1',
+                type: 'string',
+                prefix: "A:"
+              },
+                {
+                  text: 'item b',
+                  linkId: '2',
+                  type: 'string'
+                }]
+            };
+            var out = fhir.SDC.convertQuestionnaireToLForms(fhirData);
+            assert.equal(out.items[0].prefix, "A:");
+            assert.equal(out.items[1].prefix, undefined);
+          });
+
           it('should convert FHTData to lforms', function () {
             var fhtClone = angular.copy(FHTData);
             createLinkId(fhtClone.items);
@@ -369,7 +398,7 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
             convertedLfData = new LForms.LFormsData(convertedLfData);
             var reConvertedFhirQ = LForms.Util.getFormFHIRData('Questionnaire', fhirVersion, convertedLfData);
             assert.deepEqual(reConvertedFhirQ, fhirQ);
-  
+
             assert.equal(convertedLfData.name, 'USSG-FHT, (with mock-up items for skip logic demo)');
             assert.equal(convertedLfData.code, '54127-6N');
             assert.equal(convertedLfData.codeSystem, 'LOINC');
@@ -699,6 +728,18 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
 
           });
 
+          it('should covert a prefix of an item', function () {
+            var item = {
+              "questionCode": "12345",
+              "prefix": "A:",
+              "question": "fill in weight",
+              "dataType": "ST"
+            };
+            var out = fhir.SDC._processItem(item, {});
+            assert.equal(out.prefix, "A:");
+            assert.equal(out.text, "fill in weight");
+          });
+
           if(fhirVersion === 'STU3') {
             describe('argonaut samples', function () {
               it('should parse housing', function (done) {
@@ -891,7 +932,6 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
             });
           }).done().fail(function(err){console.log(': Unable to load ' + qFile);});
         });
-
       });
     });
   })(fhirVersions[i]);