+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [18.0.0] 2019-09-12
+### Added
+- Added support for 'prefix' in FHIR Questionnaire and in LForms.
+### Fixed
+- Fixed a bug that on CWE items, user typed answers that are not in the answer list are 
+  not displayed in the "OTHER" fields, when answer list is displayed as radio buttons or 
+  checkboxes.
+- Fixed a bug that code system of an answer is not imported for QuestionnaireResponse and
+  DiagnosticReport.
+### Changed
+- User typed, not-on-list value for CWE item is now stored as a string.   
+
 ## [17.4.1] 2019-09-06
 ### Fixed
 - Fixed a bug in conversion of restrictions to corresponding FHIR extensions.